this is an updated branch
i added some code, this version is not the oop version made by nigel, so everything is still in main method. but it is ok to not apply oop in this mp cuz it does not completely have attributes and behaviors, since simulation sya.

but it can further be optimized by including more methods within the class that are used in the main method.